==> 1 JSON to CSV Extraction.ipynb <==
import os
import numpy as np
import pandas as pd
import json
import csv

path_to_json = r'D:\論文\PHEME9\Data\prince-toronto/'
folders = [pos_json for pos_json in os.listdir(path_to_json) ]
base = r'D:\論文\PHEME9\Data\prince-toronto'
# to display full text
pd.set_option('display.max_colwidth', -1)

#within_tweetfolders = [pos for pos in os.listdir(r'D:\論文\PHEME9\Data\sydneysiege\rumours\544268186292867072')]

for folder in folders:
    
    if (folder == 'non-rumours' or folder == 'rumours'):
            within_tweetid = os.path.join(base, folder)
            #print(within_tweetid)
            within_folders = [pos for pos in os.listdir(within_tweetid)]
            #print(within_folders)
            for file in within_folders:
                if not file.startswith('.'):
                    
                    src_df = []
                    within_tweetfolder = os.path.join(base, folder, file)
                    #print (within_tweetfolder)

                    src_tweet = []
                    for pos in os.listdir(within_tweetfolder):
                        if not pos.startswith('.') | pos.endswith('.json'):
                            src_tweet.append(pos)
                    
                    #print(each_tweet)
                    if src_tweet[1] == 'source-tweets':
                        source = os.path.join(base, folder, file, within_tweetfolder,'source-tweets')
                        #print(source)
                        tweets = [pos for pos in os.listdir(source)]
                        #print(tweets)
                        for tweet in tweets:
                            if not tweet.startswith('.'):
                                
                                ids = os.path.join(base, folder, file, within_tweetfolder, 'source-tweets', tweet)
                                if ids.endswith('.json'):
                                    data = []
                                    #print(ids)
                                    with open(ids,'r') as f:
                                        src_json = json.load(f)
                                    src_df = pd.DataFrame(src_json)
                                    #extracting mentions id

                                    #extracting user's info
                                    usr_prop1 = src_df.user.id
                                    #make new datframe by taking only few attributes from original src_df
                                    src_df = src_df[['id','text', 'created_at' ]]
                                    src_df['user_id'] = usr_prop1

                                    #reset index of dataframe
                                    src_df = src_df.reset_index()
                                    src_df.drop(['index'], axis = 1, inplace = True)
                                    src_df = src_df.iloc[0,:]
                                    src_df = src_df.reset_index()
                                    src_df.drop("index", axis = 1, inplace = True)

                                    #adding row by row data in csv
                                    data.append(src_df.iloc[0,0])
                                    data.append(src_df.iloc[3,0])
                                    data.append(src_df.iloc[1,0])
                                    data.append(src_df.iloc[2,0])



                    if src_tweet[0] == 'reactions':
                        reactions = os.path.join(base, folder, file, within_tweetfolder,'reactions')
                        tweets_reactions = [pos for pos in os.listdir(reactions)]
                        
                        if len(tweets_reactions) ==0:
                            for i in range(4):
                                data.append(np.nan)
                            if (folder == 'non-rumours'):
                                data.append(0)
                            elif (folder == 'rumours'):
                                data.append(1)
                            with open(r'D:\論文\PHEME9\Data\CSV\prince-toronto.csv', 'a',newline = '', encoding = 'utf-8') as csvFile:
                                writer = csv.writer(csvFile)
                                writer.writerow(data)
                            csvFile.close()
                            
                            
                        elif len(tweets_reactions) >0:
                            for tweet_reaction in tweets_reactions:
                                if not tweet_reaction.startswith('.'):
                                    mix_data = []
                                    mix_data.extend(data)
                                    ids_reactions = os.path.join(base, folder, file, within_tweetfolder, 'reactions', tweet_reaction)
                                    if ids_reactions.endswith('.json'):
                                        data_reactions = []
                                        with open(ids_reactions,'r') as f:
                                            reactions_json = json.load(f)
                                        reactions_df = pd.DataFrame(reactions_json)
                                         #extracting mentions id

                                        #extracting user's info
                                        reply_usr_prop1 = reactions_df.user.id
                                        reactions_df = reactions_df[['id','text', 'created_at']]
                                        reactions_df['user_id'] = reply_usr_prop1


                                        reactions_df = reactions_df.reset_index()
                                        reactions_df.drop(['index'], axis = 1, inplace = True)
                                        reactions_df = reactions_df.iloc[0,:]
                                        reactions_df = reactions_df.reset_index()
                                        reactions_df.drop("index", axis = 1, inplace = True)
                                        data_reactions.append(reactions_df.iloc[0,0])
                                        data_reactions.append(reactions_df.iloc[3,0])
                                        data_reactions.append(reactions_df.iloc[1,0])
                                        data_reactions.append(reactions_df.iloc[2,0])


                                        mix_data.extend(data_reactions)
                                        #assign labels
                                        if (folder == 'non-rumours'):
                                            mix_data.append(0)
                                        elif (folder == 'rumours'):
                                            mix_data.append(1)
                                        #data_reactions.append(0)
                                        with open(r'D:\論文\PHEME9\Data\CSV\prince-toronto.csv', 'a',newline = '', encoding = 'utf-8') as csvFile:
                                            writer = csv.writer(csvFile)
                                            writer.writerow(mix_data)
                                        csvFile.close()





exact_data = pd.read_csv(r'D:\論文\PHEME9\Data\CSV\prince-toronto.csv', encoding = 'utf-8', header = None)

exact_data.shape

exact_data.shape
