==> 8 Chi-squared-all.ipynb <==
# Two-category variables Chi-squared test on the whole dataset.

## 0. Load data

import numpy
import pandas as pd

import os
path = r'D:\論文\PHEME9\Data\CSV'
import pathlib
data_temp = sorted(pathlib.Path(path).glob('*-2.csv'))

i = 0
all_data = pd.DataFrame()
size = 0 

while i <len(data_temp):
    rep = pd.read_csv(data_temp[i],encoding = 'utf-8', header = 0)
    src = pd.read_csv(data_temp[i+1], encoding = 'utf-8', header =0)
    df_merge = pd.merge(rep, src[['src_tweet_id','cleaned_src_tw','src_sentiment','src_emo_ro']],\
               on='src_tweet_id')
    length = rep.shape[0]
    size = size +length
    print(size)
    all_data = all_data.append(df_merge)
    i = i+2

all_data

all_data.columns

### form dual sentiment and dual emotion

dual_sentiment = all_data[['src_sentiment', 'reply_sentiment']].apply(tuple,axis=1).astype('category')

dual_sentiment.value_counts()

dual_emotion = all_data[['src_emo_ro', 'reply_emo_ro']].apply(tuple,axis=1).astype('category')
dual_emotion.value_counts()

## 1. Sentiment chi-squared test

from scipy.stats import chi2_contingency,chi2

table_se = pd.crosstab(all_data.label, dual_sentiment, rownames=['rumor classification'],colnames=['sentiment'])

stat, p, dof, expected = chi2_contingency(table_se)

# interpret test-statistic
prob = 0.95
critical = chi2.ppf(prob, dof)
if abs(stat) >= critical:
    print('Dependent (reject H0)')
    print('p-value: {}'.format(p))
    print('statisitcs: {}; critical value: {}'.format(stat,critical))
else:
    print('Independent (fail to reject H0)')
    print('statisitcs: {}; critical value: {}'.format(stat,critical))

## 2. Emotion chi-squared test

table_em = pd.crosstab(all_data.label, dual_emotion, rownames=['rumor classification'],colnames=['emotion'])

table_em

stat, p, dof, expected = chi2_contingency(table_em)

# interpret test-statistic
prob = 0.95
critical = chi2.ppf(prob, dof)
if abs(stat) >= critical:
    print('Dependent (reject H0)')
    print('p-value: {:.3e}'.format(p))
    print('statisitcs: {}; critical value: {}'.format(stat,critical))

else:
    print('Independent (fail to reject H0)')
    print('statisitcs: {}; critical value: {}'.format(stat,critical))
