==> 10 Target.ipynb <==
# XXX - Target group

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import os.path
from nltk import FreqDist, word_tokenize, corpus
import gensim
from gensim import corpora,models
from wordcloud import WordCloud
import pyLDAvis.gensim_models
from pprint import pprint

df = pd.read_csv(r'D:\論文\PHEME9\Data\CSV\charliehebdo-df-2.csv', \
                 encoding = 'utf-8', header = 0)
src_tw_df = pd.read_csv(r'D:\論文\PHEME9\Data\CSV\charliehebdo-src-2.csv', \
                        encoding = 'utf-8', header =0)

len(src_tw_df.loc[src_tw_df.label==0])+len(df.loc[df.label==0]) #number of rumorous tweets

rumor_tw = df.cleaned_reply_tw[df.label ==0 ].append(src_tw_df.cleaned_src_tw[src_tw_df.label ==0])

non_rumor_tw = df.cleaned_reply_tw[df.label ==1 ].append(src_tw_df.cleaned_src_tw[src_tw_df.label ==1])

rumor_tw_list = rumor_tw.astype(str).apply(word_tokenize).tolist()
non_rumor_tw_list = non_rumor_tw.astype(str).apply(word_tokenize).tolist()

# Rumor LDA model

# texts = rumor_tw_list+ non_rumor_tw_list

#Rumor dictionary
id2word = corpora.Dictionary(rumor_tw_list)
print(id2word)

# Term Document Frequency
corpus_r = [id2word.doc2bow(tw) for tw in rumor_tw_list]

 # Build LDA model
lda_model_r = gensim.models.ldamodel.LdaModel(corpus=corpus_r,id2word=id2word,num_topics=5, random_state=100,
                                            update_every=1,chunksize=100,passes=10,alpha='auto',per_word_topics=True)

## Rumor topics 

rumor_topics = pyLDAvis.gensim_models.prepare(lda_model_r,corpus_r,id2word)

pyLDAvis.display(rumor_topics)

p_r = pyLDAvis.gensim_models.prepare(lda_model_r, corpus_r, id2word)

save_path = 'D:/論文/PHEME9/Code/graph/charliehebdo\\'
pyLDAvis.save_html(p_r, save_path+'lda_r.html')

# Print the Keyword in the 5 topics
pprint(lda_model_r.print_topics())

# Non-rumor LDA model

id2word = corpora.Dictionary(non_rumor_tw_list)
corpus_nr = [id2word.doc2bow(tw) for tw in non_rumor_tw_list]

lda_model_nr = gensim.models.ldamodel.LdaModel(corpus=corpus_nr,id2word=id2word,num_topics=5, random_state=100,
                                            update_every=1,chunksize=100,passes=10,alpha='auto',per_word_topics=True)

## Non_rumor topics 

non_rumor_topics = pyLDAvis.gensim_models.prepare(lda_model_nr,corpus_nr,id2word)

pyLDAvis.display(non_rumor_topics)

pprint(lda_model_nr.print_topics())

p_nr = pyLDAvis.gensim_models.prepare(lda_model_nr, corpus_nr, id2word)

save_path = 'D:/論文/PHEME9/Code/graph/charliehebdo\\'
pyLDAvis.save_html(p_nr, save_path+'lda_nr.html')

# Rumor TF-IDF

## gensim_models: TfidfModel

model_tfidf_r = models.TfidfModel(corpus_r)

tfidf_r = model_tfidf_r[corpus_r]
print(tfidf_r[0])

td = {}
for document in tfidf_r:
    for token_id, score in document:
        current_score = td.get(id2word.get(token_id), 0)
        if current_score < score:
            td.update([(id2word.get(token_id), score)])

# # Sort the items of ``td`` into a new variable ``sorted_td``
# # the ``reverse`` starts from highest to lowest
# sorted_td = sorted(td.items(), key=lambda kv: kv[1], reverse=True) 

# for term, weight in sorted_td[:50]: # Print the top 50 terms in the entire corpus
#     print(term, weight)

## sklearn package: TfidfVectorizer

from sklearn.feature_extraction.text import TfidfVectorizer
vectorizer = TfidfVectorizer(max_features = 100)
X = vectorizer.fit_transform(rumor_tw.astype(str).tolist())
terms_r = vectorizer.get_feature_names()

sums = X.sum(axis=0)

data = []
for col, term in enumerate(terms_r):
    data.append( (term, sums[0,col] ))

ranking = pd.DataFrame(data, columns=['term','rank']).sort_values('rank', ascending=False)
ranking[:50]

# Non-rumor TF-IDF

vectorizer = TfidfVectorizer(max_features = 100)
Y = vectorizer.fit_transform(non_rumor_tw.astype(str).tolist())
terms_n = vectorizer.get_feature_names()

sums_n = Y.sum(axis=0)
d = []
for col, term in enumerate(terms_n):
    d.append( (term, sums_n[0,col] ))

ranking = pd.DataFrame(d, columns=['term','rank']).sort_values('rank', ascending=False)
ranking[:50]

# Common words (tf-idf)

vectorizer = TfidfVectorizer(max_features = 100)
Z = vectorizer.fit_transform(rumor_tw.append(non_rumor_tw).astype(str).tolist())
terms = vectorizer.get_feature_names()
sums = Z.sum(axis=0)


temp = []
for col, term in enumerate(terms):
    temp.append( (term, sums[0,col] ))

ranking = pd.DataFrame(temp, columns=['term','rank']).sort_values('rank', ascending=False)
ranking[:50]
