==> 12 Killing events.ipynb <==
# Targeting group among rumor tweets in five killing events 

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from nltk import word_tokenize
import seaborn as sns

import os
path = r'D:\論文\PHEME9\Data\CSV'
import pathlib
# data_temp = sorted(pathlib.Path(path).glob('*-2.csv'))

killing_events = ['charliehebdo','ferguson','germanwings','sydneysiege','ottawashooting']
killing_events.sort()
killing_events

killing_data = pd.DataFrame()
col_names = ['tweet_id','user_id','date','src_tweet_id','rumor_label','original_tweet','cleaned_tweet',
               'sentiment','emotion','source_reply','event']

# merge 5 killing events into one dataframe: killing_data

for i in killing_events:
    rep_path = pathlib.Path(path, i+'-df-2.csv')
    src_path = pathlib.Path(path, i+'-src-2.csv')
    rep = pd.read_csv(rep_path, encoding = 'utf-8', header = 0)
    src = pd.read_csv(src_path, encoding = 'utf-8', header =0)
    
    rep = rep[['reply_tweet_id', 'reply_user_id','reply_date','src_tweet_id', 'label', 'reply_tweet',
                         'cleaned_reply_tw', 'reply_sentiment', 'reply_emo_ro']]
    rep['source_reply'] = 'reply'
    rep['event'] = i
    
    rep.columns = col_names
    
    src = src[['src_tweet_id', 'src_user_id', 'src_date','src_tweet_id', 'label','src_tweet', 
                         'cleaned_src_tw','src_sentiment','src_emo_ro']]
    src['source_reply'] = 'source'
    src['event'] = i
    src.columns = col_names
    
    killing_data = pd.concat([killing_data,rep,src])

    print(killing_data.shape[0])

#  rumor tweets in five killing events : rumor_data
rumor_data = killing_data[killing_data.rumor_label==1]
rumor_data.reset_index(inplace=True,drop=True)

rumor_data.date = pd.to_datetime(rumor_data.date)
rumor_data.tweet_id = rumor_data.tweet_id.astype(str)
rumor_data.user_id = rumor_data.user_id.astype(str)

rumor_data.isnull().sum()

rumor_data.dropna(inplace=True)
rumor_data.reset_index(drop=True,inplace=True)

rumor_data.shape

rumor_data.head(1)

28905 rumorous tweets in five killing events

# Label tweets which targeted any group

'''
Updated keyword list file with filename: '12targetgroup_words.text'
read into python : target_list
'''

with open('12targetgroup_words.txt') as f:
    lines = f.readlines()
target_list = [line.strip() for line in lines]

print(target_list)
print("\n# of keywords: ", len(target_list))

'''
Create a binary variable into rumor_data dataframe - 'target_label' : 
    1, if cleaned tweets contained any keyword in the pre-defined keyword list; 
    0, otherwise.
'''

rumor_data['target_label'] = rumor_data.cleaned_tweet.astype(str).apply(lambda x: 1 
                                                                        if any(word in target_list for word in x.split()) 
                                                                        else 0 )

pd.pivot_table(rumor_data, values='tweet_id', index='event', columns='target_label',
               aggfunc='count', margins = True)

table = pd.pivot_table(rumor_data, index=['event', 'source_reply'],columns='target_label',values='tweet_id',aggfunc='count',margins=True)
# table['target_0'] = (table.C / table.groupby(level=0).C.transform(sum) * 100).astype(str) + '%'
# print table
table.columns = ['no_target','target','total']
table

table.pipe(lambda d: d.div(d['total'], axis='index')).applymap('{:.2%}'.format)

#pd.set_option('display.max_colwidth', None)
# rumor_data.loc[0,['event','target_label','original_tweet']]    

## Save rumor_data to "targetgroup.csv" file. 

**targetgroup.csv** is a dataset which contains only rumorous tweets and their replies from five killing events(Charlie Hebdo, Ferguson, Germanwings, Ottawashooting, Sydney Siege), and was labled if targeting groups or not. 

rumor_data.to_csv(r'D:\論文\PHEME9\Data\CSV\targetgroup.csv',index=False)

# Sample from each events and Save to 5 csv files

event_size = rumor_data.groupby(['event']).size().to_list()
event_size

'''
Sample 100 for 5 events and save them as 5 csv files. 
'''
j = 0
for j in range(5):
    #random sampling 50 from each target_label (either ==0 or ==1) 
    nontarget_data = rumor_data.loc[(rumor_data.event==killing_events[j])&(rumor_data.target_label ==0 )].sample(n=50,random_state = 30)
    target_data = rumor_data.loc[(rumor_data.event==killing_events[j])&(rumor_data.target_label ==1 )].sample(n=50,random_state = 30)
    sampling_data = pd.concat([nontarget_data, target_data], ignore_index=True)
    
#     path_targetsamples = r'D:\論文\PHEME9\Data\CSV\targetsamples_1012'
#     path_save = pathlib.Path(path_targetsamples, killing_events[j]+'-targetsample.csv')
#     sampling_data.to_csv(path_save,index=False)
    
    print(sampling_data.value_counts(['target_label']))
    print('event: ',killing_events[j])
    j +=1

'''
# random sample 10 tweets from each event.
sample_index = []
i = 0
for size in event_size:

    # random sample 10 tweets from each event. 
    a = list(np.random.randint(i,i+size-1,10))
    sample_index.extend(a)
    i += size
    print('event size: ',size)
    #print(i)
len(sample_index)
'''

## sentiment analysis

1. When rumorous tweets targeted the Muslims/Terrorist group, they tended to have negative sentiment. 
2. The sentiment among those rumorous tweets which didn't explicitly target any group was almost evenly distributed.

g = rumor_data.groupby(['target_label'])['sentiment'].value_counts(normalize=True).mul(100).rename('count').reset_index()\
.pipe((sns.barplot,'data'), x ='target_label',y='count',hue='sentiment',palette="Set2")

g.axes.set_ylim(0,100)
g.axes.set_title('Sentiment Analysis of Rumor Tweet w.r.t (Non-/)Targeted group')
g.axes.set_xticklabels(['Non-Targeted','Targeted'])
g.axes.set_ylabel('Count (%)')

for p in g.axes.patches:
    txt = str(p.get_height().round(1)) + '%'
    txt_x = p.get_x() 
    txt_y = p.get_height()
    g.axes.text(txt_x,txt_y,txt)
# plt.savefig('D:/論文/PHEME9/Code/graph/targetgroup-sentiment.png',dpi=300)

rumor_data[rumor_data.target_label==1].groupby(['event'])['sentiment'].value_counts(normalize=True).to_frame()

rumor_data[rumor_data.target_label==1].groupby(['event'])['sentiment'].value_counts(normalize=True).mul(100).rename('count').reset_index()\
.pipe((sns.lineplot,'data'), x ='event',y='count',hue='sentiment',palette="Set2")
