==> 17 inter_annotator_agreement.ipynb <==
# inter-annotator agreement

Manual check by Yu-Chen & Shakshi. 
Data stored in the folder: /1125 models/17_data/*sample_labeled.csv*

from sklearn.metrics import cohen_kappa_score, confusion_matrix
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

label_check = pd.read_csv('sydneysiege/17 saved_models/sample_model_targetlabel-2.csv',header=0,encoding='utf-8')

label_check

confusion_matrix(label_check['Yu-Chen'],label_check['Shakshi'])

cohen_kappa_score(label_check['Yu-Chen'],label_check['Shakshi'])

## Model vs Annotators

confusion_matrix(label_check['Shakshi'],label_check['targetlabel'])

confusion_matrix(label_check['Yu-Chen'],label_check['targetlabel'])

cohen_kappa_score(label_check['Yu-Chen'],label_check['targetlabel'])

cohen_kappa_score(label_check['Shakshi'],label_check['targetlabel'])
