==> 18 sentiment-replytarget.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from datetime import datetime ,time

# rumor reply tweets
reply = pd.read_csv(r"D:\論文\PHEME9\Code\sydneysiege\18 targetanalysis\targetdata\sydneysiege_target_reply_entire.csv", encoding = 'utf-8', header = 0,index_col=0)
reply.head(1)

print(reply.shape)
print(reply.isnull().sum())

reply.targetlabel.value_counts()

reply.reply_date = pd.to_datetime(reply.reply_date)

reply.sort_values(by='reply_date',inplace=True)
reply.reset_index(inplace=True,drop=True)
reply

target_index = reply.loc[reply.targetlabel ==1].index
print("number of targeted source tweets: ",len(target_index))
# target_index

# Calculate sentiment changes before target

# create "before" dictionary to store sentiments of non-targeting replys,
# where has three values: ratios of three sentiments/ negative to positive ratio / 
# and indices of those non-targeting replys
j = 0
before = {"sentiment": [],"np_ratio" :[],"index":[]}
target = list()
for i in target_index:
  if j == i: #the next source tweet also had target label == 1 
    before['sentiment'].append(np.nan)
    before['np_ratio'].append(np.nan)
    before['index'].append(np.nan)
    b = reply.loc[i,'reply_sentiment']
    target.append(b)
    
  else:     
    count = reply.loc[j:i-1,'reply_sentiment'].value_counts(normalize=True,sort=False)
    # ratio of three sentiments
    sen = count.round(2).to_dict()
    before['sentiment'].append(sen)
    #ratio of Negative to Positive
    try:
      neg = count['Negative']
      pos = count['Positive']
    except:
      if 'Positive' in count.index:
        # some positive tws but no negative ones
        n_p = 0
      elif "Negative" in count.index:
        # some negative tws but no positive ones
        n_p = np.inf
      else:
        # only neutral tws
        n_p = 0 
    else:
      n_p = round(neg/pos,2)
    finally:
      before['np_ratio'].append(n_p)
    before['index'].append([*range(j,i)])
    # target's sentiment
    b = reply.loc[i,'reply_sentiment']
    target.append(b)

  j =i+1

pd.set_option('display.max_colwidth', None)
sentiment_change = pd.DataFrame({"sentiment_beforetarget" : before['sentiment'],
                                 "neg_to_pos_beforetarget" : before['np_ratio'],
                                 "sentiment_target" : target,
                                 "target_index": target_index})


# Calculate sentiment changes after target arrived until the next one came.

# calculte the sentiment changes after the "last" targeing reply arrive.
# save it into "last_value" variable.

sentiment_aftertarget = sentiment_change['sentiment_beforetarget']
last_value = reply.loc[target_index[-1]+1: ,'reply_sentiment'].value_counts(normalize=True,sort=False).round(2).to_dict()
sentiment_aftertarget = pd.concat([sentiment_aftertarget[1:] ,pd.Series([last_value])], ignore_index=True )
sentiment_aftertarget

sentiment_change['sentiment_aftertarget'] = sentiment_aftertarget
sentiment_change

try:
    neg = last_value['Negative']
    pos = last_value['Positive']
except:
  if 'Positive' in last_value.keys():
    n_p = 0
  elif "Negative" in last_value.keys():
    n_p = np.inf
  else:
    n_p = 0
else:
  n_p = round(neg/pos,2)



# neg = last_value['Negative']
# pos = last_value['Positive'] 
# n_p = round(neg/pos,2)

neg_to_pos_aftertarget = sentiment_change['neg_to_pos_beforetarget'][1:]
neg_to_pos_aftertarget = pd.concat([neg_to_pos_aftertarget, pd.Series([n_p])], ignore_index=True )

sentiment_change['neg_to_pos_aftertarget'] = neg_to_pos_aftertarget
sentiment_change

## calculate number of rows

# calculate number of rows before the target sentiment arrived.
row_before = []
for i in range(len(target_index)):
  if i >0 :
    row_before.append(target_index[i] - target_index[i-1]-1)
  else:
    row_before.append(target_index[i])
print(row_before)


# calculate number of rows after the target sentiment arrived.
row_after = []
for i in range(len(target_index)):
  if i+1 < len(target_index):
    row_after.append(target_index[i+1] - target_index[i]-1)
  else: #last one
    row_after.append(reply.index[-1]-target_index[i])
print(row_after)


sentiment_change['count_of_rows_bef'] = pd.Series(row_before)
sentiment_change['count_of_rows_after'] = pd.Series(row_after)

sentiment_change.columns

sentiment_change = sentiment_change[['sentiment_beforetarget', 'neg_to_pos_beforetarget','count_of_rows_bef',  
                                    'sentiment_target',
                                    'sentiment_aftertarget', 'neg_to_pos_aftertarget','count_of_rows_after']]
sentiment_change

# save file

savepath = r'sydneysiege\18 targetanalysis\twanalysis\reply_target_sentiment.csv'

with open(savepath, 'w', encoding = 'utf-8') as f:
  sentiment_change.to_csv(f)
