==> 19 Combine targetdata.ipynb <==
Output: (save to folder '/18 analysis')
1. ***'sydneysiege_target_reply_entire.csv'*** (rumor reply data)
2. ***'combined_rumor_target.csv'*** (rumor src+reply data)

# Load library

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import os

os.chdir('D:\\論文\\PHEME9\\Code')

# Combine rumor reply data 
1. from manual-targetlabel 
2. from model-targetlabel 

then, store the combined data into the variable: *rumor_reply*

reply_model_target = pd.read_csv('sydneysiege/18 targetanalysis/targetdata/reply_model_targetlabel.csv',header=0,encoding='utf-8',index_col=0)
reply_model_target.shape

## manual-targetlabel

reply_manual_target = pd.read_csv('sydneysiege/18 targetanalysis/targetdata/sydneysiege-target-reply-600.csv",
                                  header = 0,encoding='utf-8')
reply_manual_target.shape

## concat these two dataframe

set(reply_manual_target.columns) == set(reply_model_target.columns)

rumor_reply = pd.concat([reply_model_target,reply_manual_target],axis = 0,ignore_index = True)

# check if there is any duplicates
rumor_reply.duplicated(subset='reply_tweet_id').sum()

rumor_reply.isnull().sum()

rumor_reply.reset_index(inplace=True,drop=True)
rumor_reply

rumor_reply.shape

## extract other information from original data

steps:
1. load entire reply data (rumor and nonrumor reply tweets) variable: "*reply*"
2. extract some columns from *reply* dataframe and merge them with *rumor_reply* on reply_tweet_id

### step1

# 1. Load entire reply data
reply = pd.read_csv(r'D:\論文\PHEME9\Data\CSV\sydneysiege-df.csv', encoding = 'utf-8', header = 0)
reply.head()

reply.columns

### step2

2. Extract other information from original rumor reply data *(reply)*.
***rumor_reply*** dataframe

- 'src_user_id'
- 'src_tweet'
- 'src_date'

left merge on reply_tweet_id

temp_df = reply.loc[reply.reply_tweet_id.isin(list(rumor_reply.reply_tweet_id)),
          ['reply_tweet_id','src_user_id','src_tweet','src_date']]
temp_df.shape

#### merge dataframe

then store the data into the variable: 
**rumor_reply_entire**

rumor_reply_entire = rumor_reply.merge(temp_df,on='reply_tweet_id',how = 'left')

rumor_reply_entire.shape

## save rumor reply data

with open('sydneysiege/18 targetanalysis/targetdata/sydneysiege_target_reply_entire.csv', 'w', encoding = 'utf-8') as f:
  rumor_reply_entire.to_csv(f)

# Combine source tws & Reply tws

#Load rumor reply tweet data (rumor_reply_entire)
rumor_reply_entire = pd.read_csv('sydneysiege/18 targetanalysis/targetdata/sydneysiege_target_reply_entire.csv',
                                 encoding='utf-8',header=0,index_col=0)
rumor_reply_entire.head()

temp_rumor = rumor_reply_entire[['reply_tweet_id', 'reply_user_id','reply_date', 'rumorlabel',
                                 'reply_tweet','cleaned_reply_tw', 'targetlabel', 'reply_sentiment']]

temp_rumor.columns = ['tweet_id','user_id','date','rumorlabel','tweet','cleaned_text','targetlabel','sentiment']                               

temp_rumor['src_reply'] = pd.Series(['reply'] * temp_rumor.shape[0])
temp_rumor.shape

# Load rumor source tweet data
rumor_src = pd.read_csv('sydneysiege/18 targetanalysis/targetdata/sydneysiege-target-src.csv',
                                 encoding='utf-8',header=0)

rumor_src.isnull().sum()

rumor_src.columns

rumor_src.columns = ['tweet_id','user_id','date','rumorlabel','tweet','cleaned_text','targetlabel','sentiment']

rumor_src['src_reply'] = pd.Series(['src'] * rumor_src.shape[0])

# Combine two dataframes
combined_data = pd.concat([temp_rumor,rumor_src],axis=0,ignore_index=True)
combined_data.shape

## save rumor data (src+reply)

with open('sydneysiege/18 targetanalysis/targetdata/combined_rumor_target.csv', 'w', encoding = 'utf-8') as f:
  combined_data.to_csv(f)
